discord-bot/requirements.txt
discord.py>=2.3,<3
feedparser>=6.0,<7
